chore: remove old toolchain files and consolidate into project root
Drop Jest, TSLint, Rollup configs, all tsconfig variants, the old
package-lock.json, .editorconfig, and root .gitignore. Merge package
metadata into the Vite+ package.json. Move LICENSE, README, and all
scaffold contents from file-selector/ to the project root.
