--- .editorconfig
@@ -1,18 +1,0 @@
-# EditorConfig helps developers define and maintain consistent coding styles between different editors and IDEs
-# http://editorconfig.org
-
-root = true
-
-[*]
-indent_style = space
-indent_size = 4
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*{.yml}]
-indent_size = 2
-
-[*.md]
-trim_trailing_whitespace = false
-max_line_length = off
