+++ .github/workflows/release.yml
@@ -14,10 +14,10 @@ jobs:
     name: Release
     runs-on: ubuntu-latest
     permissions:
-      contents: write      # to be able to publish a GitHub release
-      issues: write        # to be able to comment on released issues
+      contents: write # to be able to publish a GitHub release
+      issues: write # to be able to comment on released issues
       pull-requests: write # to be able to comment on released pull requests
-      id-token: write      # to enable use of OIDC for npm provenance
+      id-token: write # to enable use of OIDC for npm provenance
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
