+++ .github/workflows/stale.yml
@@ -1,24 +1,24 @@
 name: "Close stale issues and PRs"
 on:
   schedule:
-  - cron: "0 0 * * *"
+    - cron: "0 0 * * *"
 
 jobs:
   stale:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/stale@v3
-      with:
-        repo-token: ${{ secrets.GITHUB_TOKEN }}
-        days-before-stale: 60
-        days-before-close: 7
-        stale-issue-message: >
+      - uses: actions/stale@v3
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          days-before-stale: 60
+          days-before-close: 7
+          stale-issue-message: >
             This issue has been automatically marked as stale because it has not had
             recent activity in the past 60 days. It will be closed within 7 days if no further activity occurs. If the issue persists please comment here to bump your issue.
             Thank You - React Dropzone Maintaners
-        stale-pr-message: >
+          stale-pr-message: >
             This PR has been automatically marked as stale because it has not had
             recent activity in the past 60 days. It will be closed within 7 days if no further activity occurs. If the issue persists please comment here to bump your issue.
             Thank You - React Dropzone Maintaners
-        stale-issue-label: stale
-        stale-pr-label: stale
+          stale-issue-label: stale
+          stale-pr-label: stale
