+++ .gitignore
@@ -1,35 +1,7 @@
-# See http://help.github.com/ignore-files/ for more about ignoring files.
-
-## Node
-# Logs
-*.log
-npm-debug.log*
-
-# Dependency directory
-# https://docs.npmjs.com/misc/faq#should-i-check-my-node-modules-folder-into-git
 node_modules
-
-
-## OS X
-.DS_Store
-.AppleDouble
-.LSOverride
-
-# Icon must end with two \r
-Icon
-
-# Thumbnails
-._*
-
-
-## Direnv
-.envrc
-
-
-## Dist
-coverage
-build
-docs
 dist
-out-tsc
-tmp
+*.log
+.DS_Store
+!.vscode/
+!.vscode/settings.json
+!.vscode/extensions.json
