--- file-selector/.gitignore
@@ -1,7 +1,0 @@
-node_modules
-dist
-*.log
-.DS_Store
-!.vscode/
-!.vscode/settings.json
-!.vscode/extensions.json
