--- file-selector/.vscode/extensions.json
@@ -1,3 +1,0 @@
-{
-    "recommendations": ["VoidZero.vite-plus-extension-pack"]
-}
