--- file-selector/.vscode/settings.json
@@ -1,22 +1,0 @@
-{
-    "npm.scriptRunner": "vp",
-    "editor.defaultFormatter": "oxc.oxc-vscode",
-    "[javascript]": {
-        "editor.defaultFormatter": "oxc.oxc-vscode"
-    },
-    "[javascriptreact]": {
-        "editor.defaultFormatter": "oxc.oxc-vscode"
-    },
-    "[typescript]": {
-        "editor.defaultFormatter": "oxc.oxc-vscode"
-    },
-    "[typescriptreact]": {
-        "editor.defaultFormatter": "oxc.oxc-vscode"
-    },
-    "oxc.fmt.configPath": "./vite.config.ts",
-    "editor.formatOnSave": true,
-    "editor.formatOnSaveMode": "file",
-    "editor.codeActionsOnSave": {
-        "source.fixAll.oxc": "explicit"
-    }
-}
