--- file-selector/package.json
@@ -1,50 +1,0 @@
-{
-    "name": "file-selector",
-    "version": "0.0.0",
-    "description": "A starter for creating a TypeScript package.",
-    "homepage": "https://github.com/author/library#readme",
-    "bugs": {
-        "url": "https://github.com/author/library/issues"
-    },
-    "license": "MIT",
-    "author": "Author Name <author.name@mail.com>",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/author/library.git"
-    },
-    "files": [
-        "dist"
-    ],
-    "type": "module",
-    "exports": {
-        ".": "./dist/index.mjs",
-        "./package.json": "./package.json"
-    },
-    "publishConfig": {
-        "access": "public"
-    },
-    "scripts": {
-        "build": "vp pack",
-        "dev": "vp pack --watch",
-        "test": "vp test",
-        "check": "vp check",
-        "prepublishOnly": "vp run build"
-    },
-    "devDependencies": {
-        "@types/node": "^25.6.2",
-        "@typescript/native-preview": "7.0.0-dev.20260509.2",
-        "bumpp": "^11.1.0",
-        "typescript": "^6.0.3",
-        "vite-plus": "^0.1.20"
-    },
-    "overrides": {
-        "vite": "npm:@voidzero-dev/vite-plus-core@0.2.4"
-    },
-    "devEngines": {
-        "packageManager": {
-            "name": "npm",
-            "version": "12.0.1",
-            "onFail": "download"
-        }
-    }
-}
