--- file-selector/tsconfig.json
@@ -1,20 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "esnext",
-        "lib": ["es2023", "dom"],
-        "moduleDetection": "force",
-        "module": "nodenext",
-        "moduleResolution": "nodenext",
-        "resolveJsonModule": true,
-        "types": ["node"],
-        "strict": true,
-        "noUnusedLocals": true,
-        "declaration": true,
-        "noEmit": true,
-        "allowImportingTsExtensions": true,
-        "esModuleInterop": true,
-        "isolatedModules": true,
-        "verbatimModuleSyntax": true,
-        "skipLibCheck": true
-    }
-}
