--- jest.config.js
@@ -1,15 +1,0 @@
-/** @type {import('ts-jest').ProjectConfigTsJest} **/
-module.exports = {
-    testEnvironment: 'jsdom',
-    "setupFilesAfterEnv": [
-      "<rootDir>/test-setup.js"
-    ],
-    transform: {
-      '^.+.tsx?$': [
-        'ts-jest',
-        {
-          tsconfig: 'tsconfig.spec.json',
-        },
-      ],
-    },
-};
