+++ package.json
@@ -2,22 +2,17 @@
   "name": "file-selector",
   "version": "0.0.0-development",
   "description": "Convert DataTransfer object to a list of File objects",
-  "main": "./dist/index.js",
-  "module": "./dist/es2015/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist/**/*",
-    "src/*",
-    "!*.spec.*"
-  ],
   "keywords": [
-    "drag-and-drop",
-    "html5",
-    "file-api",
     "DataTransfer",
-    "File"
+    "File",
+    "drag-and-drop",
+    "file-api",
+    "html5"
   ],
   "homepage": "https://github.com/react-dropzone/file-selector",
+  "bugs": {
+    "url": "https://github.com/react-dropzone/file-selector/issues"
+  },
   "license": "MIT",
   "author": {
     "name": "Roland Groza",
@@ -27,43 +22,39 @@
     "type": "git",
     "url": "https://github.com/react-dropzone/file-selector.git"
   },
+  "files": [
+    "dist"
+  ],
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.mjs",
+    "./package.json": "./package.json"
+  },
   "publishConfig": {
     "access": "public"
   },
   "scripts": {
-    "prebuild": "npm run clean",
-    "build": "npm-run-all -s compile build:umd",
-    "build:umd": "rollup -c ./rollup.config.mjs",
-    "compile": "npm-run-all -p compile:es2015 compile:cjs compile:types",
-    "compile:es2015": "tsc -p ./tsconfig.es2015.json",
-    "compile:cjs": "tsc -p ./tsconfig.cjs.json",
-    "compile:types": "tsc -p ./tsconfig.types.json",
-    "clean": "rm -rf dist/*",
-    "lint": "tslint -c tslint.json -p ./tsconfig.spec.json -t stylish",
-    "lint:fix": "npm run lint -- --fix",
-    "pretest:cov": "npm run lint",
-    "test:cov": "jest --coverage",
-    "test": "jest --watch"
-  },
-  "dependencies": {
-    "tslib": "^2.7.0"
+    "build": "vp pack",
+    "dev": "vp pack --watch",
+    "test": "vp test",
+    "check": "vp check",
+    "prepublishOnly": "vp run build"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^28.0.0",
-    "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-terser": "^0.4.4",
-    "@types/jest": "^29.5.13",
-    "@types/node": "^22.7.5",
-    "camelcase": "^8.0.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "npm-run-all": "^4.1.5",
-    "rollup": "^4.24.0",
-    "ts-jest": "^29.2.5",
-    "tslint": "^6.1.3",
-    "typescript": "^5.6.2"
+    "@types/node": "^25.6.2",
+    "@typescript/native-preview": "7.0.0-dev.20260509.2",
+    "jsdom": "^29.1.1",
+    "typescript": "^6.0.3",
+    "vite-plus": "^0.1.20"
   },
-  "engines": {
-    "node": ">= 12"
+  "overrides": {
+    "vite": "npm:@voidzero-dev/vite-plus-core@0.2.4"
+  },
+  "devEngines": {
+    "packageManager": {
+      "name": "npm",
+      "version": "12.0.1",
+      "onFail": "download"
+    }
   }
 }
