--- test-setup.js
@@ -1,6 +1,0 @@
-// NOTE: Let us test {isSecureContext}!
-Object.defineProperty(globalThis, "isSecureContext", {
-  value: true,
-  writable: true,
-  enumerable: true,
-});
