--- tsconfig.cjs.json
@@ -1,8 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es5",
-        "outDir": "./dist"
-    }
-}
