--- tsconfig.es2015.json
@@ -1,10 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "es2015",
-        "moduleResolution": "node",
-        "target": "es2015",
-        "outDir": "./dist/es2015",
-        "importHelpers": true
-    }
-}
