--- tsconfig.spec.json
@@ -1,12 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es5",
-        "suppressImplicitAnyIndexErrors": true,
-        "esModuleInterop": true
-    },
-    "include": [
-        "src/**/*.ts"
-    ]
-}
