--- tsconfig.types.json
@@ -1,11 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "declaration": true,
-        "emitDeclarationOnly": true,
-        "declarationDir": "./dist",
-        "moduleResolution": "node",
-        "module": "es2015",
-        "target": "esnext"
-    }
-}
